requirements.txt
Flask
torch
torchvision
tensorflow
numpy
requests
pillow